+++ src/compas_fab/robots/state.py
@@ -45,12 +45,12 @@ class RobotCellState(Data):
     @property
     def tool_ids(self):
         # type: () -> List[str]
-        return self.tool_states.keys()
+        return list(self.tool_states.keys())
 
     @property
     def rigid_body_ids(self):
         # type: () -> List[str]
-        return self.rigid_body_states.keys()
+        return list(self.rigid_body_states.keys())
 
     @property
     def __data__(self):
@@ -60,6 +60,37 @@ class RobotCellState(Data):
             "tool_states": self.tool_states,
             "rigid_body_states": self.rigid_body_states,
         }
+
+    def __eq__(self, value):
+        # type: (RobotCellState) -> bool
+        if value is None or not isinstance(value, RobotCellState):
+            return False
+        if self.robot_flange_frame != value.robot_flange_frame:
+            return False
+        if self.robot_configuration:
+            try:
+                if not self.robot_configuration.close_to(value.robot_configuration):
+                    return False
+            except:
+                return False
+        else:
+            if self.robot_configuration != value.robot_configuration:
+                return False
+        if set(self.tool_states.keys()) != set(value.tool_states.keys()):
+            return False
+        for tool_id, tool_state in self.tool_states.items():
+            if tool_state is None and value.tool_states[tool_id] is None:
+                continue
+            if tool_state != value.tool_states[tool_id]:
+                return False
+        if set(self.rigid_body_states.keys()) != set(value.rigid_body_states.keys()):
+            return False
+        for rigid_body_id, rigid_body_state in self.rigid_body_states.items():
+            if rigid_body_state is None and value.rigid_body_states[rigid_body_id] is None:
+                continue
+            if rigid_body_state != value.rigid_body_states[rigid_body_id]:
+                return False
+        return True
 
     @classmethod
     def from_robot_cell(cls, robot_cell, robot_configuration=None):
@@ -139,7 +170,7 @@ class RobotCellState(Data):
         """
         tool_id = self.get_attached_tool_id(group)
         if not tool_id:
-            return None
+            return []
         ids = []
         for rigid_body_id, rigid_body_state in self.rigid_body_states.items():
             if rigid_body_state.attached_to_tool == tool_id:
@@ -147,10 +178,8 @@ class RobotCellState(Data):
         return ids
 
     def get_attached_rigid_body_ids(self):
-        # type: (str) -> List[str]
-        """Returns the ids of the rigid bodies attached to the robot
-
-        This does not include the tools attached to the robot and the workpieces attached to the tools.
+        # type: () -> List[str]
+        """Returns the ids of the rigid bodies attached to links of the robot and to tools.
 
         Returns
         -------
@@ -159,12 +188,15 @@ class RobotCellState(Data):
         """
         ids = []
         for rigid_body_id, rigid_body_state in self.rigid_body_states.items():
-            if rigid_body_state.attached_to_link:
+            if rigid_body_state.attached_to_link or rigid_body_state.attached_to_tool:
                 ids.append(rigid_body_id)
+        return ids
 
-    def set_tool_attached_to_group(self, tool_id, group, attachment_frame=None, touch_links=None, detach_others=True):
-        # type: (str, str, Optional[Frame], Optional[List[str]], Optional[bool]) -> None
+    def set_tool_attached_to_group(self, tool_id, group, attachment_frame=None, touch_links=None):
+        # type: (str, str, Optional[Frame], Optional[List[str]]) -> None
         """Sets the tool attached to the planning group.
+
+        Any other tools that is attached to the specified planning group will be detached.
 
         Notes
         -----
@@ -181,8 +213,7 @@ class RobotCellState(Data):
             Defaults to None, which means that the tool's frame coincides with the end frame of the planning group.
         touch_links : list of str, optional
             The names of the robot links that are allowed to collide with the tool.
-        detach_others : bool, optional
-            Whether to detach all other tools from the group. Defaults to True.
+
         """
         if not attachment_frame:
             attachment_frame = Frame.worldXY()
@@ -192,10 +223,40 @@ class RobotCellState(Data):
         self.tool_states[tool_id].attachment_frame = attachment_frame
         self.tool_states[tool_id].touch_links = touch_links or []
 
-        if detach_others:
-            for id, tool_state in self.tool_states.items():
-                if id != tool_id and tool_state.attached_to_group == group:
-                    tool_state.attached_to_group = None
+        # Detach other tools that are attached to the same group
+        for id, tool_state in self.tool_states.items():
+            if id != tool_id and tool_state.attached_to_group == group:
+                self.set_tool_detached(id)
+
+    def set_tool_detached(self, tool_id, frame=None, touch_links=None):
+        # type: (str, Optional[Frame], Optional[List[str]]) -> None
+        """Sets the tool to be detached from the planning group.
+
+        Parameters
+        ----------
+        tool_id : str
+            The id of the tool.
+        frame : :class:`compas.geometry.Frame`, optional
+            The frame of the tool (relative to the world coordinate frame) after detaching.
+            Defaults to None, which means that the tool's frame is not changed.
+        touch_links : list of str, optional
+            The names of the robot links that are allowed to collide with the tool.
+            Defaults to None, which means that the tool's touch links are reset to an empty list.
+            If this behavior is not desired, consider modifying the tool_states directly
+            or set the touch_links set explicitly.
+        """
+        self.tool_states[tool_id].attached_to_group = None
+
+        if frame:
+            self.tool_states[tool_id].frame = frame
+        else:
+            # If frame is not specified, the frame of the tool is not changed after detaching
+            pass
+
+        if touch_links is None:
+            self.tool_states[tool_id].touch_links = []
+        else:
+            self.tool_states[tool_id].touch_links = touch_links
 
     def set_rigid_body_attached_to_link(self, rigid_body_id, link_name, attachment_frame=None, touch_links=None):
         # type: (str, str, Optional[Frame | Transformation], Optional[List[str]]) -> None
@@ -272,13 +333,13 @@ class ToolState(Data):
 
     Attributes
     ----------
-    frame : :class:`compas.geometry.Frame`
+    frame : :class:`compas.geometry.Frame` | :class:`compas.geometry.Transformation`
         The base frame of the tool relative to the world coordinate frame.
         If the tool is attached to a planning group, this frame can be set to None.'
         In that case, the planner or visualization tool will use the end frame of the planning group.
     attached_to_group : :obj:`str`, optional
         The name of the robot planning group to which the tool is attached. Defaults to ``None``.
-    attachment_frame : :class:`compas.geometry.Frame`, optional
+    attachment_frame : :class:`compas.geometry.Frame` | :class:`compas.geometry.Transformation`, optional
         The frame of the tool relative to the frame of the attached link. Defaults to ``None``.
     touch_links : :obj:`list` of :obj:`str`
         The names of the robot links that are allowed to collide with the tool.
@@ -298,12 +359,18 @@ class ToolState(Data):
         configuration=None,
         is_hidden=False,
     ):
-        # type: (Frame, Optional[str], Optional[Frame], Optional[List[str]], Optional[Configuration], Optional[bool]) -> None
+        # type: (Frame | Transformation, Optional[str], Optional[Frame| Transformation], Optional[List[str]], Optional[Configuration], Optional[bool]) -> None
         super(ToolState, self).__init__()
         self.frame = frame  # type: Frame
+        # Convert frame to a Frame if it is a Transformation
+        if isinstance(frame, Transformation):
+            self.frame = Frame.from_transformation(frame)
         self.attached_to_group = attached_to_group  # type: Optional[str]
         self.touch_links = touch_links or []  # type: List[str]
         self.attachment_frame = attachment_frame  # type: Optional[Frame]
+        # Convert frame to a Frame if it is a Transformation
+        if isinstance(attachment_frame, Transformation):
+            self.attachment_frame = Frame.from_transformation(attachment_frame)
         self.configuration = configuration  # type: Optional[Configuration]
         self.is_hidden = is_hidden  # type: bool
 
@@ -317,6 +384,36 @@ class ToolState(Data):
             "configuration": self.configuration,
             "is_hidden": self.is_hidden,
         }
+
+    def __eq__(self, value):
+        # type: (ToolState) -> bool
+        if value is None:
+            return False
+        # Wrap the tests in try block, if any of the attributes cannot be compared,
+        # it will raise exception and return False
+        try:
+            if self.frame != value.frame:
+                return False
+            if self.attached_to_group != value.attached_to_group:
+                return False
+            if set(self.touch_links) != set(value.touch_links):
+                return False
+            if self.attachment_frame != value.attachment_frame:
+                return False
+            if self.configuration:
+                try:
+                    if not self.configuration.close_to(value.configuration):
+                        return False
+                except:
+                    return False
+            else:
+                if self.configuration != value.configuration:
+                    return False
+            if self.is_hidden != value.is_hidden:
+                return False
+        except:
+            return False
+        return True
 
 
 class RigidBodyState(Data):
@@ -379,7 +476,7 @@ class RigidBodyState(Data):
 
     Parameters
     ----------
-    frame : :class:`compas.geometry.Frame`
+    frame : :class:`compas.geometry.Frame` | :class:`compas.geometry.Transformation`
         The base frame of the rigid body relative to the world coordinate frame.
     attached_to_link : :obj:`str` , optional
         The name of the robot link to which the rigid body is attached.
@@ -417,6 +514,9 @@ class RigidBodyState(Data):
         # type: (Frame, Optional[str], Optional[str], Optional[List[str]], Optional[List[str]], Optional[Frame], Optional[bool]) -> None
         super(RigidBodyState, self).__init__()
         self.frame = frame  # type: Frame
+        # Convert frame to a Frame if it is a Transformation
+        if isinstance(frame, Transformation):
+            self.frame = Frame.from_transformation(frame)
         self.attached_to_link = attached_to_link  # type: Optional[str]
         self.attached_to_tool = attached_to_tool  # type: Optional[str]
         self.touch_links = touch_links or []  # type: List[str]
@@ -440,3 +540,28 @@ class RigidBodyState(Data):
             "attachment_frame": self.attachment_frame,
             "is_hidden": self.is_hidden,
         }
+
+    def __eq__(self, value):
+        # type: (RigidBodyState) -> bool
+        if value is None:
+            return False
+        # Wrap the tests in try block, if any of the attributes cannot be compared,
+        # it will raise exception and return False
+        try:
+            if self.frame != value.frame:
+                return False
+            if self.attached_to_link != value.attached_to_link:
+                return False
+            if self.attached_to_tool != value.attached_to_tool:
+                return False
+            if set(self.touch_links) != set(value.touch_links):
+                return False
+            if set(self.touch_bodies) != set(value.touch_bodies):
+                return False
+            if self.attachment_frame != value.attachment_frame:
+                return False
+            if self.is_hidden != value.is_hidden:
+                return False
+        except:
+            return False
+        return True